#1519 Moved loop swap transformation to proper location.
