+++ src/psyclone/psyir/transformations/__init__.py
@@ -68,6 +68,8 @@ from psyclone.psyir.transformations.fold_conditional_return_expressions_trans \
 from psyclone.psyir.transformations.transformation_error \
     import TransformationError
 from psyclone.psyir.transformations.omp_taskwait_trans import OMPTaskwaitTrans
+from psyclone.psyir.transformations.loop_swap_trans import LoopSwapTrans
+
 # The entities in the __all__ list are made available to import directly from
 # this package e.g.:
 # from psyclone.psyir.transformations import ExtractTrans
@@ -80,6 +82,7 @@ __all__ = ['ChunkLoopTrans',
            'ReadOnlyVerifyTrans',
            'RegionTrans',
            'Abs2CodeTrans',
+           'LoopSwapTrans',
            'LoopTrans',
            'LoopFuseTrans',
            'Matmul2CodeTrans',
